// FrontEnd/ClientApp/src/models/teamStatistics.ts
export class TeamStatistics {
  points: number = 0;
  pointsHome: number = 0;
  pointsAway: number = 0;

  games: number = 0;
  gamesHome: number = 0;
  gamesAway: number = 0;

  wins: number = 0;
  winsHome: number = 0;
  winsAway: number = 0;

  draws: number = 0;
  drawsHome: number = 0;
  drawsAway: number = 0;

  losts: number = 0;
  lostsHome: number = 0;
  lostsAway: number = 0;

  scoredGols: number = 0;
  scoredGolsHome: number = 0;
  scoredGolsAway: number = 0;

  concededGols: number = 0;
  concededGolsHome: number = 0;
  concededGolsAway: number = 0;
}